requirements.txt
fastapi
uvicorn[standard]
databricks-sdk
pandas
openpyxl
snowflake-connector-python
